requirements.txt
websockets
rich
bittensor
bittensor-cli
python-telegram-bot>=20.0
pyyaml
